Add support for public key algo 20 (Elgamal Encrypt and Sign)
There are still older public keys out in the wild that use this older and
by my understanding deprecated public key algorithm specifier. So far as I can
tell from the older and newer versions of the RFC, this code (20) was used to
specify Elgamal with "usage flags" of Encrypt and Sign. Code 16 was there as
well but specified Elgamal for encryption only.

Being as though both codes use Elgamal, this patch simply provides pass-through
for this older code (20) all the places in the code where code 16 was being
checked.